Implement `EndpointMethodGenerator`
